Project Demo
Project working demonstration. The server script is running on Raspberry Pi 400 and the Mobile App is running in a virtual environment. The Pi and App are communication over a local Wi-Fi network. The Pi is interfaced with a doorbell hardware. The doorbell housing is 3D printed and it contains camera, mic, speaker, LCD display, PIR motion sensor and a DC buzzer. The door lock is controlled by using a DC solenoid.